Make SHM implementation switchable to `shared_memory` crate
