+++ ipc_test/Cargo.toml
@@ -18,3 +18,6 @@ raw_sync = "0.1.5"
 sendfd = "0.4.3"
 serde = { version = "1.0.149", features = ["derive"] }
 log = "0.4.17"
+nix = "0.26.2"
+tempfile = "3.3.0"
+shared_memory = "0.12.4"
