+++ libertem_asi_tpx3/src/lib.rs
@@ -4,7 +4,6 @@ pub mod main_py;
 pub mod exceptions;
 pub mod chunk_stack;
 pub mod receiver;
-pub mod shm;
 pub mod stats;
 pub mod stream;
 pub mod headers;
